Move construction of card data to separate function
